Updated xserver and mesalib to git version 23 Oct 2016
xserver          commit 5dcb0666b82f5ab00f3d22e86f05ac14b0d5341e
mesa             commit 7b7eb7170d16ddb0963900ccf59b39956219373c

+++ mesalib/src/mesa/state_tracker/st_cb_clear.c
@@ -318,8 +318,8 @@ is_scissor_enabled(struct gl_context *ctx, struct gl_renderbuffer *rb)
    return (ctx->Scissor.EnableFlags & 1) &&
           (scissor->X > 0 ||
            scissor->Y > 0 ||
-           scissor->X + scissor->Width < rb->Width ||
-           scissor->Y + scissor->Height < rb->Height);
+           scissor->X + scissor->Width < (int)rb->Width ||
+           scissor->Y + scissor->Height < (int)rb->Height);
 }
 
 /**
